+++ app/src/androidTest/java/com/github/HumanLearning2021/HumanLearningApp/view/dataset_editing/DatasetsOverviewActivityTest.kt
@@ -1,4 +1,4 @@
-package com.github.HumanLearning2021.HumanLearningApp
+package com.github.HumanLearning2021.HumanLearningApp.view.dataset_editing
 
 import android.content.Intent
 import androidx.test.core.app.ApplicationProvider
@@ -10,6 +10,8 @@ import androidx.test.espresso.intent.matcher.IntentMatchers.hasComponent
 import androidx.test.espresso.matcher.ViewMatchers.*
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.github.HumanLearning2021.HumanLearningApp.R
+import com.github.HumanLearning2021.HumanLearningApp.TestUtils
 import com.schibsted.spain.barista.assertion.BaristaVisibilityAssertions.assertDisplayed
 import dagger.hilt.android.testing.HiltAndroidRule
 import dagger.hilt.android.testing.HiltAndroidTest
@@ -23,7 +25,7 @@ import org.junit.Before
 @RunWith(AndroidJUnit4::class)
 
 @HiltAndroidTest
-class DataOverviewActivityTest {
+class DatasetsOverviewActivityTest {
 
     @get:Rule
     var hiltRule = HiltAndroidRule(this)
@@ -31,7 +33,7 @@ class DataOverviewActivityTest {
     val activityScenarioRule: ActivityScenarioRule<AddPictureActivity> = ActivityScenarioRule(
         Intent(
             ApplicationProvider.getApplicationContext(),
-            DataOverviewActivity::class.java
+            DatasetsOverviewActivity::class.java
         )
     )
 
@@ -58,7 +60,7 @@ class DataOverviewActivityTest {
     @Test
     fun rightActivityIsStartedAfterCreateButton() {
         onView(withId(R.id.dataOverviewButton)).perform(click())
-        intended(hasComponent(DataCreationActivity::class.java.name))
+        intended(hasComponent(CategoriesEditingActivity::class.java.name))
     }
