+++ app/src/main/AndroidManifest.xml
@@ -15,18 +15,18 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.HumanLearning">
-        <activity android:name=".view.DisplayImageSetActivity" />
-        <activity android:name=".DataOverviewActivity" />
+        <activity android:name=".view.dataset_editing.DisplayImageSetActivity" />
+        <activity android:name=".view.dataset_editing.DatasetsOverviewActivity" />
         <activity
-            android:name=".view.LearningDatasetSelectionActivity"
+            android:name=".view.learning.LearningDatasetSelectionActivity"
             android:label="@string/LearningDatasetSelection_title" />
-        <activity android:name=".DataCreationActivity" />
-        <activity android:name=".AddPictureActivity" />
-        <activity android:name=".view.DisplayImageActivity" />
-        <activity android:name=".view.DisplayDatasetActivity" />
-        <activity android:name=".view.LearningSettingsActivity" />
-        <activity android:name=".view.LearningActivity" />
-        <activity android:name=".MainActivity">
+        <activity android:name=".view.dataset_editing.CategoriesEditingActivity" />
+        <activity android:name=".view.dataset_editing.AddPictureActivity" />
+        <activity android:name=".view.dataset_editing.DisplayImageActivity" />
+        <activity android:name=".view.dataset_editing.DisplayDatasetActivity" />
+        <activity android:name=".view.learning.LearningSettingsActivity" />
+        <activity android:name=".view.learning.LearningActivity" />
+        <activity android:name=".view.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
