+++ app/src/main/java/com/github/HumanLearning2021/HumanLearningApp/presenter/LearningPresenter.kt
@@ -4,7 +4,7 @@ import android.app.Activity
 import android.widget.ImageView
 import com.github.HumanLearning2021.HumanLearningApp.hilt.Demo2Database
 import com.github.HumanLearning2021.HumanLearningApp.model.*
-import com.github.HumanLearning2021.HumanLearningApp.view.LearningMode
+import com.github.HumanLearning2021.HumanLearningApp.view.learning.LearningMode
 import javax.inject.Inject
