+++ app/src/main/java/com/github/HumanLearning2021/HumanLearningApp/view/dataset_editing/DatasetsOverviewActivity.kt
@@ -1,31 +1,31 @@
-package com.github.HumanLearning2021.HumanLearningApp
+package com.github.HumanLearning2021.HumanLearningApp.view.dataset_editing
 
 import android.content.Intent
 import android.os.Bundle
-import android.os.PersistableBundle
 import android.util.Log
 import android.view.View
 import androidx.appcompat.app.AppCompatActivity
-import com.github.HumanLearning2021.HumanLearningApp.view.fragments.DatasetListFragment
+import com.github.HumanLearning2021.HumanLearningApp.R
+import com.github.HumanLearning2021.HumanLearningApp.view.dataset_list_fragment.DatasetListFragment
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class DataOverviewActivity : AppCompatActivity() {
+class DatasetsOverviewActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_data_overview)
+        setContentView(R.layout.activity_datasets_overview)
 
         val dsListFragment = supportFragmentManager.findFragmentById(R.id.dataOverview_fragment)
         if (dsListFragment is DatasetListFragment) {
             dsListFragment.selectedDataset.observe(this) {
                 Log.d("DataOverview activity", "selected ds is  $it")
                 // TODO include Dataset (it) with Intent
-                startActivity(Intent(this, DataCreationActivity::class.java))
+                startActivity(Intent(this, CategoriesEditingActivity::class.java))
             }
         }
     }
 
     fun launchDataCreationActivity(view: View){
-        startActivity(Intent(this, DataCreationActivity::class.java))
+        startActivity(Intent(this, CategoriesEditingActivity::class.java))
     }
 }
