+++ app/src/main/java/com/github/HumanLearning2021/HumanLearningApp/view/dataset_list_fragment/DatasetListFragment.kt
@@ -1,7 +1,6 @@
-package com.github.HumanLearning2021.HumanLearningApp.view.fragments
+package com.github.HumanLearning2021.HumanLearningApp.view.dataset_list_fragment
 
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -16,7 +15,6 @@ import com.github.HumanLearning2021.HumanLearningApp.hilt.DummyDatabase
 import com.github.HumanLearning2021.HumanLearningApp.model.DatabaseManagement
 import com.github.HumanLearning2021.HumanLearningApp.model.Dataset
 import dagger.hilt.android.AndroidEntryPoint
-import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 /**
