+++ app/src/main/res/layout/activity_camera.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:layout_gravity="center_horizontal"
-    tools:context=".AddPictureActivity">
+    tools:context=".view.dataset_editing.AddPictureActivity">
 
     <RelativeLayout
         android:layout_width="match_parent"
