+++ app/src/main/res/layout/activity_display_dataset.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.DisplayDatasetActivity">
+    tools:context=".view.dataset_editing.DisplayDatasetActivity">
 
     <EditText
         android:id="@+id/display_dataset_name"
