+++ app/src/main/res/layout/activity_display_image_set.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.DisplayImageSetActivity">
+    tools:context=".view.dataset_editing.DisplayImageSetActivity">
 
     <TextView
         android:id="@+id/display_image_set_name"
