+++ app/src/main/res/layout/activity_learning.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.LearningActivity">
+    tools:context=".view.learning.LearningActivity">
 
     <ImageView
         android:id="@+id/learning_im_to_sort"
