+++ app/src/main/res/layout/activity_learning_settings.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:layout_gravity="center"
-    tools:context=".view.LearningSettingsActivity">
+    tools:context=".view.learning.LearningSettingsActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
